refactor(alert): extract shared alert node configuration into alertBase.ts and simplify Alert and AlertSchema implementations
