+++ packages/alert/src/Alert.ts
@@ -1,13 +1,11 @@
-import { mergeAttributes, Node } from '@tiptap/core';
+import { Node } from '@tiptap/core';
 import { ReactNodeViewRenderer } from '@tiptap/react';
 
+import { alertBaseConfig } from './alertBase';
+import type { AlertOptions, AlertType } from './alertBase';
 import { AlertComponent } from './AlertComponent';
 
-export type AlertType = 'info' | 'warning' | 'danger' | 'success' | 'note' | 'tip';
-
-export interface AlertOptions {
-  HTMLAttributes: Record<string, any>;
-}
+export type { AlertType, AlertOptions } from './alertBase';
 
 declare module '@tiptap/core' {
   interface Commands<ReturnType> {
@@ -29,68 +27,7 @@ declare module '@tiptap/core' {
 }
 
 export const Alert = Node.create<AlertOptions>({
-  name: 'alert',
-
-  group: 'block',
-
-  content: 'block+',
-
-  defining: true,
-
-  addOptions() {
-    return {
-      HTMLAttributes: {},
-    };
-  },
-
-  addAttributes() {
-    return {
-      type: {
-        default: 'info',
-        parseHTML: (element) => element.getAttribute('data-alert-type') || 'info',
-        renderHTML: (attributes) => {
-          return {
-            'data-alert-type': attributes.type,
-          };
-        },
-      },
-    };
-  },
-
-  parseHTML() {
-    return [
-      {
-        tag: 'div[data-type="alert"]',
-        getAttrs: (node) => {
-          if (typeof node === 'string') return false;
-
-          const element = node as HTMLElement;
-          const type = element.getAttribute('data-alert-type');
-
-          return type ? { type } : false;
-        },
-      },
-      {
-        tag: 'div[data-alert-type]',
-        getAttrs: (element) => {
-          const type = (element as HTMLElement).getAttribute('data-alert-type');
-
-          return type ? { type } : false;
-        },
-      },
-    ];
-  },
-
-  renderHTML({ node, HTMLAttributes }) {
-    return [
-      'div',
-      mergeAttributes(this.options.HTMLAttributes, HTMLAttributes, {
-        'data-type': 'alert',
-        class: `alert alert-${node.attrs.type}`,
-      }),
-      0,
-    ];
-  },
+  ...alertBaseConfig,
 
   addCommands() {
     return {
